+++ sources/samples/vhosts-http-1.1.ts
@@ -17,32 +17,12 @@
 import * as http from "../";
 
 /**
- * 创建路由器对象 A
+ * 创建路由器对象 B，虽然真实 URI 是以 /admin 开头，但是作为挂载点，前缀会被自动
+ * 去除。因此 /admin 前缀不需要在路由中填写。
  */
 let routerA = http.createStandardRouter();
 
 routerA.get("/", async function(ctx) {
-
-    const req = ctx.request;
-    const resp = ctx.response;
-
-    resp.send(`Visiting system A ${req.realPath}. (Request Path: ${req.path})`);
-
-}).get("/a", async function(ctx) {
-
-    const req = ctx.request;
-    const resp = ctx.response;
-
-    resp.send(`Visiting system A ${req.realPath}. (Request Path: ${req.path})`);
-});
-
-/**
- * 创建路由器对象 B，虽然真实 URI 是以 /admin 开头，但是作为挂载点，前缀会被自动
- * 去除。因此 /admin 前缀不需要在路由中填写。
- */
-let routerB = http.createStandardRouter();
-
-routerB.get("/", async function(ctx) {
 
     const req = ctx.request;
     const resp = ctx.response;
@@ -61,9 +41,9 @@ routerB.get("/", async function(ctx) {
  * 创建路由器对象 C，虽然真实 URI 是以 /statics 开头，但是作为挂载点，前缀会被自动
  * 去除。因此 /statics 前缀不需要在路由中填写。
  */
-let routerC = http.createStandardRouter();
+let routerB = http.createStandardRouter();
 
-routerC.get("/", async function(ctx) {
+routerB.get("/", async function(ctx) {
 
     const req = ctx.request;
     const resp = ctx.response;
@@ -78,28 +58,28 @@ routerC.get("/", async function(ctx) {
     resp.send(`Visiting system C ${req.realPath}. (Request Path: ${req.path})`);
 });
 
-let serverC = http.createServer({
-    "router": routerC
-});
-
 let serverB = http.createServer({
     "router": routerB
 });
 
+let serverA = http.createServer({
+    "router": routerA
+});
+
 /**
- * 创建一个监听 0.0.0.0:8080 端口的 HTTP 服务器，并指定使用 router 对象作为路由器。
+ * 创建一个监听 0.0.0.0:80 端口的 HTTP 服务器，并指定使用 router 对象作为路由器。
  */
-let serverA = http.createMountableServer({
-    "port": 8080,
-    "router": routerA,
-    "mounts": {
-        "/admin": serverB,
-        "/statics": serverC
+let dispatcher = http.createVirtualDispatcher({
+    "port": 80,
+    "hosts": {
+        "default": serverA,
+        "wp.local.org": serverA,
+        "m.local.org": serverB
     }
 });
 
 // 启动服务器
-serverA.start().then(() => {
+dispatcher.start().then(() => {
 
     console.log("服务器成功启动");